requirements.txt
# Requerimientos principales del proyecto
psycopg2-binary
python-dotenv
pandas
openpyxl
xlrd
gspread
google-auth

# Herramientas de desarrollo (opcionales)
pytest
mypy
black
isort